Рендер карточек фильмов по клику на кнопку
