+++ src/view/button.js
@@ -1,3 +1,3 @@
- export const createButtonLoaderTemplate = () => {
+export const createButtonLoaderTemplate = () => {
   return `<button class="films-list__show-more">Show more</button>`;
- };
+};
