+++ src/view/comment.js
@@ -13,5 +13,5 @@ export const createCommentTemplate = (comment) => {
       <button class="film-details__comment-delete">Delete</button>
     </p>
   </div>
-</li>`
+</li>`;
 };
