+++ src/view/film-details.js
@@ -1,8 +1,8 @@
 export const createfilmDetailsTemplate = (card) => {
-  const {name, img, fullDescription, comments, raiting, year, director, writers, actors, releaseDay,
+  const {name, img, fullDescription, comments, raiting, director, writers, actors, releaseDay,
     ageToWatch, runtime, country, genre, id} = card;
 
-  return `<section class="film-details">
+  return `<section class="film-details" id=${id}>
   <form class="film-details__inner" action="" method="get">
     <div class="form-details__top-container">
       <div class="film-details__close">
@@ -120,5 +120,5 @@ export const createfilmDetailsTemplate = (card) => {
       </section>
     </div>
   </form>
-</section>`
+</section>`;
 };
