+++ src/view/menu.js
@@ -8,5 +8,5 @@ export const createMenuTemplate = () => {
       <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">8</span></a>
     </div>
     <a href="#stats" class="main-navigation__additional">Stats</a>
-  </nav>`
+  </nav>`;
 };
