+++ src/view/sort.js
@@ -4,5 +4,5 @@ export const createSortTemplate = () => {
     <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
     <li><a href="#" class="sort__button">Sort by date</a></li>
     <li><a href="#" class="sort__button">Sort by rating</a></li>
-  </ul>`
+  </ul>`;
 };
