+++ src/view/stat.js
@@ -2,5 +2,5 @@ export const createStatTemplate = () => {
   return `
   <section class="footer__statistics">
     <p>130 291 movies inside</p>
-  </section>`
+  </section>`;
 };
