+++ src/view/user.js
@@ -1,8 +1,6 @@
 export const createUserTemplate = () => {
-  return `
-  <section class="header__profile profile">
+  return `<section class="header__profile profile">
     <p class="profile__rating">Movie Buff</p>
     <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-  </section>
-`
+  </section>`;
 };
